Create Uploading the data
This script downloads the data from GitHub, saves them as .json file, and uploads them to an earlier created Google Cloud bucket. 

Afterwards, it creates a BigQuery dataset. For uploading the data from Storage to BigQuery dataset a schema for the data is required. You can generate it with any tool listed here:
https://stackoverflow.com/questions/7341537/tool-to-generate-json-schema-from-json-data

With json schema saved onto your local disc, it uploads the data to BigQuery dataset and is ready for the SQL analysis.